Switch from threads to multiprocessing after doing some research about why my threading was making it slower instead of faster!
As it turns out, threads are _not_ concurrent in Python, and of course there is additional overhead for creating and managing them. So they're only useful when doing heavy I/O operations, so that while one process is waiting for an I/O operation to finish, another one will be unblocked to start another I/O operation. For this type of thing, which is really much more CPU bound than I/O bound (despite the fact that we're reading from the hard-drive, that part seems to take a negligible amount of time in comparison to the computation of filtering and testing the files.

References:
https://medium.com/practo-engineering/threading-vs-multiprocessing-in-python-7b57f224eadb
http://www.informit.com/articles/article.aspx?p=1850445&seqNum=9